# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos-CMS 后台管理API
    description: Kratos-CMS 后台管理API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-cms
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-cms/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/categories:
        get:
            tags:
                - CategoryService
            description: 获取类别列表
            operationId: CategoryService_ListCategory
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每一页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: AND过滤参数，其语法为json格式的字符串，如：{"key1":"val1","key2":"val2"}，具体请参见：https://github.com/tx7do/go-utils/tree/main/entgo/query/README.md
                  schema:
                    type: string
                - name: or
                  in: query
                  description: OR过滤参数，语法同AND过滤参数。
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件，其语法为JSON字符串，例如：{"val1", "-val2"}。字段名前加'-'为降序，否则为升序。
                  schema:
                    type: array
                    items:
                        type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListCategoryResponse'
        post:
            tags:
                - CategoryService
            description: 创建类别
            operationId: CategoryService_CreateCategory
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCategoryRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Category'
    /admin/v1/categories/{id}:
        get:
            tags:
                - CategoryService
            description: 获取类别数据
            operationId: CategoryService_GetCategory
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: fieldMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Category'
        put:
            tags:
                - CategoryService
            description: 更新类别
            operationId: CategoryService_UpdateCategory
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Category'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Category'
        delete:
            tags:
                - CategoryService
            description: 删除类别
            operationId: CategoryService_DeleteCategory
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        Category:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                parentId:
                    type: integer
                    format: uint32
                name:
                    type: string
                slug:
                    type: string
                description:
                    type: string
                thumbnail:
                    type: string
                password:
                    type: string
                fullPath:
                    type: string
                priority:
                    type: integer
                    format: int32
                createTime:
                    type: string
                updateTime:
                    type: string
                postCount:
                    type: integer
                    format: uint32
            description: 类别
...
        CreateCategoryRequest:
            type: object
            properties:
                category:
                    $ref: '#/components/schemas/Category'
                operatorId:
                    type: integer
                    format: uint32
            description: 请求 - 创建类别
...
        ListCategoryResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Category'
                total:
                    type: integer
                    format: int32
            description: 回应 - 类别列表
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...